+++ test/graphql/types/person_search_type.js
@@ -1,0 +1,7 @@
+export default `
+    union personSearch = Client | Vendor
+
+    type personQuery {
+        firstSearchResult: personSearch
+    }
+`;