+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - mergeTypes parses interface type
+- mergeTypes parses union type
 
 ## [0.0.12] - 2017-05-04
 ### Fixed